ellipsis-minor review addressed
